backend/requirements/base.txt
django>=5.0,<7.0
whitenoise>=6.5
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
Pillow>=10.0
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
channels>=4.0
channels-redis>=4.2
stripe>=8.0
google-auth>=2.0
google-auth-oauthlib>=1.0
